Extracts "system" class to its own files
